// src/app/challenge/enums/socket.enum.ts
export enum SocketEventOut {
    GET_ONLINE_USERS = 'GET_ONLINE_USERS',
    APPROVE_CHALLENGE_REQUEST = 'APPROVE_CHALLENGE_REQUEST',
    CHALLENGE_REQUEST = 'CHALLENGE_REQUEST',
}


export enum SocketEventIn {
    NEW_CHALLENGE_REQUEST = 'NEW_CHALLENGE_REQUEST',
    NEW_USER_CONNECTION = 'NEW_USER_CONNECTION',
    DISCONNECT_USER = 'DISCONNECT_USER',
    START_GAME = 'START_GAME',
    ONLINE_USERS_LIST = 'ONLINE_USERS_LIST',
    SERVER_SOCKET_ERROR = 'SERVER_SOCKET_ERROR',
    CONNECT_ERROR = 'connect_error',
    ERROR = 'error',
}